Weaviate vector database/requirements.txt
weaviate-client>=4.9.0
python-dotenv>=1.0.0
